moved global test stuff to dedicated file
